Data.Vector.Algorithms.Radix

Description

This module provides a radix sort for a subclass of unboxed arrays. The radix class gives information on * the number of passes needed for the data type

the size of the auxiliary arrays

how to compute the pass-k radix of a value

Radix sort is not a comparison sort, so it is able to achieve O(n) run time, though it also uses O(n) auxiliary space. In addition, there is a constant space overhead of 2*size*sizeOf(Int) for the sort, so it is not advisable to use this sort for large numbers of very small arrays.

A standard example (upon which one could base their own Radix instance) is Word32:

We choose to sort on r = 8 bits at a time

A Word32 has b = 32 bits total

Thus, b/r = 4 passes are required, 2^r = 256 elements are needed in an auxiliary array, and the radix function is:

radix k e = (e `shiftR` (k*8)) .&. 255

Documentation

sort :: forall e m v. (PrimMonad m, MVector v e, Radix e) => v (PrimState m) e -> m ()

Sorts an array based on the Radix instance.

sortBy

Arguments

:: (PrimMonad m, MVector v e)

=> Int
the number of passes

-> Int
the size of auxiliary arrays

-> (Int -> e -> Int)
the radix function

-> v (PrimState m) e
the array to be sorted

-> m ()

Radix sorts an array using custom radix information requires the number of passes to fully sort the array, the size of of auxiliary arrays necessary (should be one greater than the maximum value returned by the radix function), and a radix function, which takes the pass and an element, and returns the relevant radix.

class Radix e where

Minimal complete definition

passes, size, radix

Methods

passes :: e -> Int

The number of passes necessary to sort an array of es

size :: e -> Int

The size of an auxiliary array

radix :: Int -> e -> Int

The radix function parameterized by the current pass